change tests and move file
